+++ app/src/main/res/layout/activity_admin.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/color_background"
-    tools:context=".admin_activity">
+    tools:context=".ui.admin_activity">
 
     <TextView
         android:id="@+id/admin_text_title"
