+++ app/src/main/res/layout/activity_employee.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/color_background"
-    tools:context=".employee_activity">
+    tools:context=".ui.employee_activity">
 
     <androidx.appcompat.widget.AppCompatButton
         android:id="@+id/employee_btn_sale"
