+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/color_background"
-    tools:context=".login_activity">
+    tools:context=".ui.login_activity">
 
     <TextView
         android:id="@+id/login_text_name"
