+++ app/src/main/res/layout/activity_stock_admin.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/color_background"
-    tools:context=".stock_admin_activity">
+    tools:context=".ui.stock_admin_activity">
 
     <ImageView
         android:id="@+id/stockadmin_btn_back"
